requirements.txt
torch
transformers
pillow
pandas
editdistance
scikit-learn
datasets
requests
